conversational-pdf-qa-with-chat-history/requirement.txt
langchain
python-dotenv 
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-google-genai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
langchain-groq
duckduckgo-search
pymupdf
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
numexpr
pytube
